Installation debugging
I just made an addendum to the README file to help with installation. Obviously, I can't really put everything I did into the file, since it was alot and also specific to my machines, but I feel that these are definitely helpful pointers for future devs who want to work on this.