Projects and Developers Page Routing + Sponsors moved to Sponsors Page
I added the projects and developers pages to the website, and I moved the sponsors content from the homepage to the sponsors page. At the moment, none of the pages are organized well in terms of UI. I'm going to be fixing this next by individualizing the content on each page.