Home Page Mostly Done + W3 CSS Removed (except for navbar files, temporarily).
The layout of the home page is essentially done. The three red buttons in the middle do not work since I'm not really sure how to safely acquire a link with our overall setup. I moved the W3 stylesheets temporarily to the navbar .hbs files so that they don't break. We can replace them with Bootstrap later.