Massive UI Update
I overhauled the home page, fixed the look for the sponsors page, added a project card UI for the projects page, and a developer card UI for the developer page. I also added whatever else needed adding to get the pages to work. The webiste, from a front-end level, looks decently presentable now.