Slay the Beast
NOTE: This has NOT been tested yet, that is for the next commit!

Broke up the massive rcos.sql into a large number of smaller files that
conform to the format used by DbMate[1], which is similar to the format
used by other migration tools with an up and down statement.

While going through I made sure to cleanup the SQL, condensing, adding
comments, rewriting foreign key statements, and adding CHECKs, and
improving the formatting.

The new SQL format is hopefully easier to understand since everything is
broken into distinct related components with a clear up and down.

Next up is testing, additional checks, and more cleanup. Then on to more
views and stored functions.

[1]: https://github.com/amacneil/dbmate