Docker Compose and Caddyfile
Wrote an intial docker-compose.yml file for future deployments, and
updated the README and roles files to go with it.

Removed the postgrest.conf file since PostgREST is now configured
through environment variables in the compose file.

Renamed the postgrest folder to "examples" since it now just contains
the client example.