WIP error handling middleware
I've made some progress. Telescope errors all get converted into  actix errors automatically which then get converted into HTTP responses, so I cannot downcast to telescope errors in the middleware. My current workaround is to serialize telescope errors as JSON and then propagate them back as responses with a custom MIME type. The MIME type acts as a signal value for them to be caught by the custom middleware where they can be rendered into the proper webpage. The current blocker is that the body of the response is an async stream of byte buffers, which is difficult to handle.